hypetown/requirements.txt
aiogram==3.15.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
APScheduler==3.10.4
python-dotenv==1.0.1
aiohttp
tonsdk==1.0.3
